api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
marshmallow==3.20.1